+++ app/src/main/java/woowacourse/payments/list/ListActivity.kt
@@ -1,6 +1,5 @@
 package woowacourse.payments.list
 
-import android.app.Activity
 import android.content.Intent
 import android.os.Bundle
 import android.widget.Toast
@@ -40,7 +39,7 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import woowacourse.payments.NewCardActivity
+import woowacourse.payments.newCard.NewCardActivity
 import woowacourse.payments.PaymentCard
 import woowacourse.payments.domain.Card
 import woowacourse.payments.ui.theme.AndroidpaymentsTheme
