+++ app/src/main/java/woowacourse/payments/newCard/NewCardTopBar.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments
+package woowacourse.payments.newCard
 
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
